feat(visitation): switch rpc method
The default odoo rpc behavior prevents unauthorized rpc and is not easily extendable without
creating a controller. As such, this commit adds work around code to treat the OWL app as a foreign
rpc client. For example, a login and pwd is required on each call. To
achieve this, an authenticateable public user is added via xml. This is
the same as the default public user, but it has known credentials. The
effect of this is somewhat like using an api key in a client side
application in the sense that the scope of access is tightly controlled.

Additionally, this commit adds axios as a dependency to prevent some
rework of implementing the json rpc protocol.