feat(visitation): prepare to deploy in odoo online
This commit is refactors the backend structures and the js io layer to account for the conditions of
creating backend structures in odoo online. Specifically, all field and model names need be prefixed
with 'x_' and mapped to the non-prefixed version in the client app.

Note: any xmlid references will need to be matched in the online
deployment to satisfy any env.ref(...) calls.