feat(visitation): allow read access on select fields for rpc user
Not a normal technique, but the front rpc user needs to read fields to fire the automated action.
Kind of a smell, but necessary because 1) the public user cant call the rpc 2) the automated action
needs to fire in the same transaction as the write call I think and 3) the trigger fields need to
read the fields model for some reason.